Add Staging Migration Hiera data
Add Hieradata that will only apply in Staging once we've started to migrate
to A. N. Other Vendor from Skyscape. We're only doing Staging here because
we'd like it to be externally accessible in much the same way that
Production and Preview are - i.e. without having to use
govuk_select_organisation.