Ensure home directories are not world-readable
Ensure that home directories are not world-readable by setting a file
mode of `0750` on each user's home directory.

We use `directory` for the ensure parameter for consistency with our
existing behaviour, since we don't currently manage the home directory
with Puppet (if a user is deleted, the home directory remains intact).

Set file permissions after the user has been created so that we can be
sure that the user exists when setting ownership of the directory, and
also to avoid Puppet's built-in user resource from possibly clobbering
our file permissions on each run.