Do not encrypt Licensify mongo data in Preview
Using a partition that is encrypted at rest on `licensify_mongo`
machines in Preview means that these machines cannot be reboot without
manual intervention to mount the encrypted partition.

Since we don't store anything sensitive in Preview and we don't
synchronise Licensify's MongoDB data from Production to Preview, disable
encryption in Preview to remove the manual overhead incurred by having
to manually mount the disk on each boot.

This also means that we can turn these machines off overnight and start
them again in the morning (to save on costs).

[1]: https://github.gds/gds/env-sync-and-backup/blob/a2bd38a2a9a34c7c8f17d9c5596cad63f648302a/config/preview.cfg#L13