[#97091450] Reduce APT update frequency
From once every Puppet run to once a day. So that it stops clashing with
`unattended-upgrades` overnight, because they both lock `dpkg` and only one
can win, causing one of the following to happen:

- `unattended-upgrades` can't get the lock and retries again the following
  hour, which leaves less time for the `unattended-reboots` window
- Puppet can't get the lock when running `apt-get update`, which causes a
  "puppet last run errors" alert in Nagios, which prevents
  `unattended-upgrades` from rebooting because there are outstanding alerts

The update resource will still get notified when we add or modify an APT
repository. I suspect the only difference we'll see is if we tell Puppet to
ensure a specific version of a package that has only recently been added to
a repo, but that seems like a minor trade-off.

Docs for the params are here:

- https://github.com/puppetlabs/puppetlabs-apt/tree/1.7.x#apt-1

I did toy with the idea of upgrading `puppetlabs/apt` at the same time but
there are a huge amount of backwards-incompatible changes that I don't have
time to test right now.