Remove service_manual app
This was added in 81df4c8c67f4e0f01ec4b89c26d0b23ed0651700 with the
intention that the service manual would be split out into its own
app, but it's clear that that won't happen.

This app is causing Puppet errors on the frontend app servers so
it's not worth keeping it around.

    Error: Could not set 'present' on ensure: No such file or directory - /etc/govuk/service-manual/env.d/PORT20150616-17416-32zc06.lock