[#97087958] Blacklist overlayfs kernel module
As a temporary mitigation for a local root exploit in Ubuntu. This is fixed
by the kernel `3.13.0-54`, but we're not able to reboot all machines at the
moment. We can remove this when we have.

I've verified that we don't appear to be using this kernel module, because
nothing has loaded it, by running the following and getting no output:

    fab production all do:'lsmod | grep overlayfs || true' --hide running,status

That also means that we don't need to unload the module when deploying this.

NB: I'm probably going to DevOps this straight into production as soon as
the PR is merged, because of the potential impact.