Fix exit code check for rkhunter passive alert
The variable `$CODE` was being assigned later than it should have
been to check the result of the rkhunter run.

This means that we ended up in the state where rkhunter exited with
a non-zero exit code because it found something which it considered
a warning, but the Icinga alert had a code of zero (OK).