Check that apps are running the right ruby version.
When a .ruby-version file is updated, a standard app reload won't pick
up the change, and the application will be left running under the old
version of ruby.  A full restart of the application is required to pick
up the new version.

This check will help to ensure that this doesn't get overlooked.