Don't warn on reboot required in the last day
This commit changes the threshold for an Icinga warning when reboots
are required from immediately after the check runs to until at least
24 hours after the packages are installed.

I don't think that issuing a warning immediately is useful. It creates
a lot of alerts which we acknowledge because most machines will
reboot automatically during the following night.

In the case where the update is urgent (eg for security reasons)
we'll hear about it through other means and will take action to
reboot. Package upgrades that require a reboot are so frequent that
it's unlikely that delaying this Icinga warning will cause us to
miss an urgent update.