Remove default for content_store::mongodb_name
Remove the default value for the MongoDB database name for safety.

I'm about to override this value for a new s_draft_content_store machine
class, and I don't want to risk the new machines using the default
database name, which would mean draft content would be inadvertently
written to the database used for published content.