Define s_draft_cache machine class
This machine class is based on the s_cache class but will be used to
serve draft content.

I have disabled traffic replay for this machine class; I'll tackle that
in a separate story.

Note that these machines will host an instance of the router-api, which
will use its own database but share the MongoDB cluster on the
router_backend machines. This is a compromise we have made to avoid the
cost of spinning up router_backend machines for draft content.