Add mongodb_name param to router_api class
Allow the name of the MongoDB database to be used by router-api to be
set as a class parameter.

Make the class parameter mandatory as I'm about set the database name to
be used by the router-api when serving draft (unpublished) content and I
don't want to risk the draft instance of router-api connecting to the
wrong database, which might occur if I used a default value and had a
typo in the Hiera data which overrides it.