Override signon location for authenticating-proxy
authenticating-proxy will be installed on the draft stack, but will share the
same signon service as the main stack. This means that we cannot rely on Plek to
correctly identify the signon location on the draft stack (it will return
https://signon.draft.preview.publishing.service.gov.uk by default). Plek allows
the location of a given service to be overridden by setting a specific
environment variable[1]. In this case, the environment variable is
PLEK_SERVICE_SIGNON_URI. We set this to the main preview signon on
preview_draft. The same will need to be done for staging and production in due
course.

[1] See http://www.rubydoc.info/gems/plek#Hacking_Plek_URLs