Configure graphite offsite backups to be weekly
We do two daily backups of graphite data in production:

- An internal backup using rsync, from graphite-1 to backup-1
- An offsite backup using duplicity, from backup-1

The graphite data is thousands of small files, currently totalling
around 190G. Both of the backups seem to be capped by an IO bottleneck.
Because both backup processes are affected by that same bottleneck,
running both processes at once massively increases the amount of time
taken to complete.

On staging where the only backup is the internal rsync, it takes around
13 hours to complete. The data set on production is slightly larger, so
I'd expect it to take around 16 hours to complete. Due to the extra
backup process however, it takes over 30 hours to complete. That means
it overlaps with the next day's backup process, adding to the problems.

Reducing the frequency of the offsite backup to be weekly instead of
daily should reduce the competition for IO enough to allow the internal
rsync backup to completed in around 16 hours.

Starting the weekly offsite backup on a Saturday morning means that any
IO bottlenecks will be resolved by the time traffic starts ramping up
again on Monday.

Ideally we'd fix this issue by getting an order of magnitude speed up in
the daily backup -- 16 hours to read, transfer and write 190G of data is
very slow. This change is a short term fix to keep the current system
going until someone has a chance to investigate a faster backup option.