Set URI for draft content-store on draft_frontend
Override the Plek URI for the draft instance of the content-store
application on the draft_frontend machine class.

Plek is used by the frontend applications to locate microservices.

We need to override the default location for content-store on the
draft_frontend machines so that applications hosted there communicate
with the content-store used for draft content, rather than the main
content-store used for published content.