Remove weekly and monthly backup transfer jobs
Remove the weekly and monthly cronjobs that copy backups to the
`backup-1` on-site server.

We don't use these cronjobs - `/etc/backup/weekly` and
`/etc/backup/monthly` are empty.

The weekly and monthly backups are copied as part of the daily cronjob.
We should probably tidy up the semantics of that, but for now just
remove the parts we don't use.