Send output from es-rotate to syslog
Send output from the `es-rotate` command to syslog rather than a file.

The file was being truncated rather than being appended to, so it was of
limited use.

Also, it's preferable to use syslog so that we can view these logs in
Kibana without having to setup a logstream resource for the log file.