Make offsite job freshness threshold conditional
We added a `weekday` parameter to run the Graphite off-site backup
weekly instead of daily:
https://github.gds/gds/puppet/pull/2947

The Icinga passive check freshness threshold was not updated to reflect
that change, meaning that we saw alerts warning that the Graphite
backup's freshness threshold had exceeded.

Make the freshness threshold conditional on whether the backup is run
weekly or daily.