Don't hide /sidekiq for apps.
This reverts 3df6b1a (#1971).  It was added when we were going to add
sidekiq-monitoring to the apps themselves.  In the end we decided to
spin up the monitoring as a separate app. Apps using sidekiq don't
expose this endpoint by default, it requires explicitly adding a route
for it, so there's no need for this to be present.

We're now experimenting with adding it to the apps, but putting it
behind signon auth, so this needs to be removed.