Run MongoDB backup once per replica set
Our MongoDB backups were running once on each member of a replica set.
Effectively, this meant that the same data set was backed up three
times.

This sometimes causes contention, meaning that some backups fail
because others are in progress and creates incomplete backups that
remain uncompressed and use high amounts of disk space.

Instead, pass the array of replica set members to the `mongodb::backup`
class and run the backup script on the first member of the replica set
only.

When the backup runs, it will select the primary member of the replica
set and create the backup from the primary's data.

To avoid complications arising from the 'internal TLD', e.g. production,
used in the FQDN of the machines, use the `hostname` Facter fact to
determine which machine to run the backups on. This means that when
specifying members of the replicaset, we have to specify them using the
hostname only, i.e. without the vDC name.

I'm exploring using [puppetdbquery][] as a more flexible solution.

[puppetdbquery]: https://github.com/dalen/puppet-puppetdbquery