Remove frequency parameter for on-site backups
The `backup::directory` defined type had a `$frequency` parameter that
could only ever be set to 'daily' because we don't support frequencies
other than daily.

Remove the unused `frequency` parameter as we don't need or use any
other frequency for copying backups to our on-site backup machine. This
also simplifies the path to the scripts that are run when these backups
take place.