Use config_hash option when setting up Jenkins
This commit moves our Jenkins configuration out of a file and in
to hieradata.

On a machine which has been already provisioned this has the
behaviour of quoting all of our config settings. For example,
`RUN_STANDALONE=true` changes to `RUN_STANDALONE="true"`. It
will leave the comments intact.

On a new machine which does not already have `/etc/default/jenkins`
setup, this change will mean that the comments in `/etc/default/jenkins`
will never be created.