Add https_port param to loadbalancer::balance
Allow the port that Nginx listens on for HTTPS requests to be
overridden in the `loadbalancer::balance` defined type.

I plan to use this so that the load-balancing vhost for the draft
instance of content-store uses port 8443 to distinguish it from the
vhost that serves requests for the instance of content-store that
serves published content.

Downstream, this will be used to implement access controls to prevent
published frontend instances from communicating with the draft
content-store, which would be bad because the draft content-store can
contain embargoed or sensitive content.

See RFC 16 for more details:
https://gov-uk.atlassian.net/wiki/display/pubplatform/RFC+16%3A+How+to+prevent+published+%28live%29+frontends+from+reading+from+the+draft+content+store