Make draft-content-store's Nginx lb vhost use 8443
Make the Nginx virtual host on the api_lb machines that load-balances
requests for the draft instance of content-store listen on port 8443
rather than 443.

Additionally, make the vhost _only_ listen on that port.

Do this to enforce separation between the draft content-store and the
published instance. This is important so that frontend applications
serving published content to the public do not inadvertently read from
the draft content-store. We use firewall rules configured on our vShield
Edge Gateway to enforce this separation by IP address.

See RFC 16 for more details:
https://gov-uk.atlassian.net/wiki/display/pubplatform/RFC+16%3A+How+to+prevent+published+%28live%29+frontends+from+reading+from+the+draft+content+store