Start nginx only after mapit is running
Our process for updating MapIt's database involves stopping nginx for the
duration of the update so that requests will be retried quickly against the
other mapit machine, so that users don't see errors. In fact, Puppet was
starting nginx early on in the run, before the new database dump was loaded
and while mapit was still stopped.

This commit inserts mapit::nginx at the end of the dependency chain so that
nginx will only be started once mapit is running. The exec to load the
database dump in s_mapit_server already notifies mapit::service, so the new
database will already have been loaded by that point.

This change removes the ~> (notification arrow, which also sends a refresh
event) between mapit::service and mapit::end; I can't see a reason for a
notification being necessary there.