Increase critical disk time threshold on backup machines
The critical alert is triggering regularly on the backup machines when
backups run. This does not seem to actually interrupt the backups from
working, so it does not seem like a critical issue.

Looking back over the graph of disk wait times on this machine over the
past 30 days, the current threshold caused the alert to trigger 15
times.

Increasing the threshold to 1000ms which would have triggered 3 times in
the same period, which seems like a more reasonable setting for the
critical level.