Make draft-router-api routable
Add a DNS entry to make the draft instance of the router-api resolve to
the DraftRouterInternal load balancer, which fronts the draft_cache
machines where the draft instance of the router-api is installed.

Note that the main instance of router-api is installed on the
router_backend machines in the Router vDC. We install the draft instance
on the draft_cache machines to avoid the additional cost of running
router_backend machines for draft functionality.