Amend Elasticsearch index rotation time
We use the cron resource to run a script - es-rotate-passive-check - which
rotates the logs index so that we have one index created per day. When the
script runs, a new index is created with a name based on the current data,
and the logs-current alias is attached to it. This then ensures that we have
indexes on a per-day basis. The script also tidies up indexes that are older
than a set threshold of fifteen days.

Since we've been shutting machines down in Preview, this script hasn't been
running because the cron resource tells it to run at 00:01, but machines are
shut down at around 19:00 and aren't started back up again until 04:40. In
Staging and Production, we still run this at 00:01 because those stacks
aren't shut down overnight.

Runs in the morning, approximately 90mins after machines have been told to boot,
in order that the new index is created for the correct day. Running the night
before will cause issues because the date won't have changed, and running any
earlier than this might cause issues if the logs-elasticsearch boxes haven't
booted and the cluster isn't healthy. This change might mean that we end up
with some logs from the preceding 90mins before this script runs in the previous
day's Elasticsearch index, but since the indexes are all concatenated for use in
Kibana, this shouldn't affect us too much.