Set PLEK_HOSTNAME_PREFIX env var on draft machines
So that Plek knows to prepend 'draft-' to the URI of draft applications,
set the `PLEK_HOSTNAME_PREFIX` environment variable on all machines
running draft applications.

`PLEK_HOSTNAME_PREFIX` is supported by Plek v1.11.0 and later.
Affected applications must use that version of Plek or later in order
for the prefix to be prepended.

We've done this so that the content-store running on draft_content_store
machines know where to find the draft instance of the router-api.