Enforce a TMPTIME of 7 in /etc/default/rcS
/etc/default/rcS is used to apply parameters governing the behaviour of
scripts which run at boot. The TMPTIME parameter is used to determine how
long files in /tmp should exist at boot. For instance, if a file has an
mtime seven days ago or less, and the TMPTIME parameter is set to 7, then
that file would not be deleted on boot. As soon as it breaches the 7 day
threshold, it would be.