Use app_domain to set Signon/Errbit URIs
Move the Hiera configuration for authenticating-proxy out of the
environment-specific Hieradata files and into `common.yaml` and use
`app_domain` to construct the URLs for Signon and Errbit.

Prior to this, the Errbit URL was set in Staging to
`errbit.staging.alphagov.co.uk` but should have been
`errbit.production.alphagov.co.uk` since Staging uses the same internal
DNS entries as Production for historical reasons.

This change reduces duplication and the likelihood that we'll make this
error again.