Parameterise hostname of the APT mirror
Add a new value to common hieradata for `apt_mirror_hostname`.

We want all environments to use the production APT mirror, so this
can't use `app_domain`.

Add parameters to each class which sets up anything to do with the
APT mirrors and use a hiera lookup in the hieradata to reduce
duplication of `alphagov.co.uk`.