Use hiera lookups for `monitoring_domain_suffix`
In preview and production this is identical to `app_domain`.

In staging, this is not set to `app_domain`.

However once we move providers and move away from alphagov.co.uk
we're going to start setting `app_domain` in staging so that it is
`staging` and not `production`, which will remove the need for this
logic to change things across environments.