Fix disabling of app services on dev VM.
For the 'manual' stanza to take effect, it has to come after the 'start
on' stanza (this fact doesn't appear to be mentioned in the
documentation, but I've verified it by manual testing on a trusty VM).
Without this, every app service will be started on boot on the dev VM.

Depending on the speed of the host machine, and the sequence of events,
this may not be noticed because /var/govuk isn't mounted immediately at
boot (it's mounted by vagrant once the machine has booted), and all
these services depend on this filesystem.  If this filesystem has been
mounted before upstart has given up starting the services, the VM will
grind to a halt as every single app is started, and the oom killer kicks
in etc.