Script to checkout repos from a file or stdin
I've done this manually a few times before, and I suspect others
have too, so it seemed worth adding as a script. Especially as
most GOV.UK teams now have a list of repos they maintain theres
an easy starting point to get the right repos.

(Longer term it would be nice if this could resolve depdencies
through `Pinfile` and check those out as well, but this is a
first scricted step).

It either takes a path to a file of repos (one per one) or reads
line by line from stdin, so it can be piped to easily.