Respect `Link` header from whitehall-admin backend
Whitehall sends a `Link` header from the [assets backend][0] to tell
clients where to find the cover sheet for a given attachment. This
header appears to be being stripped somewhere in the stack.

[0]: https://github.com/alphagov/whitehall/blob/6c72de9390adcf9b23e90bd37686ce6fb940a41d/app/controllers/attachments_controller.rb#L45

We serve our assets via Sendfile - in Nginx land this is called via
X-Accel-Redirect.

Because X-Accel-Redirect is a [redirect internally][1], Nginx strips all
headers outside of a basic set (couldn't find official docs on this, but
[this StackOverflow answer][2] seems to mirror the behaviour I saw).

[1]: http://nginx.org/en/docs/http/ngx_http_core_module.html#internal
[2]: http://stackoverflow.com/a/24509358/61435

To restore the desired behavior, we can use Nginx's [`$upstream_http_`][3]
variables to fetch the original headers from the response.

[3]: http://nginx.org/en/docs/http/ngx_http_upstream_module.html#var_upstream_http_