Correct the docs for nagios_crit in govuk::mount
While reviewing #3076 I noticed that the documentation for this module is out
of sync with the actual defaults. The critical threshold is actually 5%, while
the documentation asserts that it is 20%. A default of 20% free space would not
make sense as the warning threshold is 10% and having 20% free space is less
serious than having 10% free space.