Add a node manifest for new mapit machines
We're starting to move MapIt to new machines on Trusty, where we will
deploy it as a normal app using the Govuk::App config. This commit is the
initial configuration for the new machines to enable us to bring one up
in Preview. We will set up the application in a separate PR as part of a
later story to get the app deployed to the new machine.

The node definition is mostly copied from the existing s_mapit_server
node definition, with a few differences:

- s_mapit does not include the mapit module. This module is concerned with
  setting up things which we will get for free (and in a more up-to-date
  way) by deploying MapIt as a normal application using alphagov-deployment.
- s_mapit doesn't include the creation of the database which is in
  s_mapit_server, for the same reason: this will be handled using
  Govuk::App for MapIt.
- s_mapit does include our standard nginx module.

The mapit.yaml hieradata file is identical to mapit_server.yaml.

We're only setting up a single mapit-1 machine in Preview for now; there
will be two in every environment in the end (as we currently have for
mapit-server-*) but we'll set up the second one later on when we need it.