Template Jenkins config for Puppet
This class is an attempt manage Jenkins through Puppet and Hiera. It's
basic function is to generate the main Jenkins configuration file
(config.xml) and restart the Jenkins service if there are any changes.

 - Sets up the oauth plugin so that Jenkins can authenticate with Github,
and when a user logs in their permissions should already be set in the main
config file so one does not need to go and manually set up their user in Jenkins.
Unfortunately this has created yet another list of users to manage in Hiera.

 - Adds a bunch of standard plugin XML configuration files

I also note there seems to be quite a few variables between Preview and
Staging/Production Jenkins instances which will need to be factored in
somehow.