Remove old things from .gitignore
We don't use RVM and we shouldn't encourage its use by having it
ignored by Git.

The `pt` gem isn't in the Gemfile for this repo so it shouldn't
be ignored by Git.

People are still free to use either of these things but they
should go in their local machine's `.gitignore`.