Update Jenkins jobs for JJB
Originally when Jenkins Job Builder was created (#3001), only production
jobs were configured. To ensure that the right jobs are added into the
right environment, the "common" jobs which are shared between
environments have been put into templates with variables which are set
in hiera.

Where jobs _only_ exist in an environment they have been added into
their own directory structure which references that environment. Welcome
comments on a different approach to this.

I have also added these jobs into `vagrant.yaml` to make the Vagrant
jenkins machine more lifelike, however happy to drop this if people are
not happy with this.