Allow specifying an SSH keypair for jenkins user
We're trying to automate as much as we can when creating an environment.
Instead of having Puppet create the keypair which we then add to
GitHub Enterprise, it's easier if we create the keypair beforehand
and add it to Puppet and GitHub Enterprise at the same time.

This allows us to bring up an environment in an automated way without
needing to stop half way through to deal with SSH keys.

If an SSH keypair is provided in hieradata Puppet will use that,
otherwise it will create one as normal.