router app: Move env var to application manifest
Move the `ROUTER_MONGO_DB` environment variable, which is specific to
the router application, from the `s_draft_cache`
node manifest to the `router` application manifest, where it logically
belongs.

I have intentionally not used a parameter default so as to defend
against the database name being misconfigured and having draft and
non-draft applications talk to the wrong database.