Set titles for application env var resources
Explicitly set a resource title for the application environment variable
resources in the router application's manifest.

The intent is to reduce the likelihood that these resource titles would
conflict with another resource on the same machine, by prefixing the
resource title with the application name ('router') (referred to here
using `$title`).