Enable traffic replay for content preview (draft)
Enable HTTP traffic replay between the draft origin used for content
preview in Production and its equivalent in Staging.

The mechanism used is the same as for our main (non-draft) origin; we
replay HTTP GET, HEAD and OPTIONS requests sent to port 7999 (Varnish).

This change specifies the target of the traffic replay in Hiera rather
in the Puppet code, which the previous implementation had avoided (see
the removed comment in the `router::gor` class in this commit's diff).

While hardcoding those values made sense at the time, there's no simple
alternative to moving these to Hiera since the `draft_cache` machine
definition includes the `cache` machine definition.

I'll add a comment in a separate commit in the Hiera data as a reminder
that the target IP/hostname is not easily changed.