Add govuk::apps::mapit class
We're moving to deploying MapIt as a normal application. This change follows
on from acaa35f14fcba1bc65a9c987ba34b240890b96d2 which added a new machine for
mapit which will run in parallel with the existing mapit-server-* machines
until we've finished the migration.

Port 3108 has already been reserved in the development repo for MapIt:

    https://github.gds/gds/development/commit/e7587351f1fed1800872b036ffa467054aa26f9e

We're defining the database in the same class as the application config
because MapIt's database is on the same machine as the app. For most apps the
database needs to be created on a different machine from the application
config so needs to be in a separate class.

MapIt is a Python/Django app so the app type is 'procfile'. We'll add a
Procfile to our fork of the application repo.