Install system dependencies for MapIt
Most of these are required by MapIt's Python dependencies, GDAL in particular.

We will be installing all of MapIt's Python dependencies as part of the normal
deploy process, using pip and virtualenv, so are only defining the system
packages needed by those here.

This thread is particularly helpful for explaining the dependencies between
the system and python geo packages:

    http://gis.stackexchange.com/questions/28966/python-gdal-package-missing-header-file-when-installing-via-pip