Change PagerDuty script to `pagerduty_nagios.pl`
This script was renamed in a0a7507e9775783f281f03e713d31602582811b0
as part of the move from Nagios to Icinga, but the reference to the
script in `pager_contact.cfg.erb` wasn't changed. This means that
Icinga hasn't been contacting PagerDuty for about the last 2 years.

This script was written by PagerDuty so we should keep it with its
original name.