Allow govuk::app to define read_timeout of proxy
An app should be able to vary the length of time a request can last
before it 504s. We have a usecase where we are ingesting large chunks of
data and this can take a decent chunk of time, so we need to be able to
extend the request timeout from the hardcoded 15s to a longer time
period.