Bugfix for setting Jenkins global variables
Global environment variables were not being set which caused an issue
with deploying Jenkins while we managed the config. I have managed to
narrow it down to a slightly incorrect XML parameter `<int>`.

The `<int>` parameter specifies the number of the Jenkins global
variables that are being set. These parameters are simply key->value and
in sets of two, so if if two appropriate `<string>` parameters are
removed then this number will be amended **otherwise zero of the global
parameters will be set**.