Add a PagerDuty weekly drill
Add some cron resources and an Icinga alert which will test every
Wednesday morning whether PagerDuty is working.

Cron will remove and replace the file with enough time for the alert
to go critical and notify PagerDuty.

The file freshness threshold for the Icinga check is set to 8 days to
override the default parameters for `check_file_age` and ensure that the
alert doesn't trigger at some point in the week between Wednesday and
Wednesday.

This happens on Wednesday morning to welcome the new team to 2nd line.