Allow DB to only be connected to over SSL
In the GOV.UK infrastructure we try to have all communication traveling
over the network encrypted. The default setup of our postgresql
databases allows a client to connect with out encrypting its network
traffic, this undermines any encryption further down the line as a bad
actor can sniff or alter the data coming from the source.

This commit allows you to set a DB so that it will only accept
connections that are encrypted using SSL. This works by setting the type
to hostssl rather than host. This indicates that the rule only applies
for the host when SSL is used, see:

http://www.postgresql.org/docs/9.3/static/auth-pg-hba-conf.html