Set proxy_connect_timeout for ES load balancers
The nginx load balancers in front of elasticsearch do not currently set
a connect timeout, and therefore get the default value of 60s[1]. This
means that if one node is unavailable (due to VM pauses or other
issues), then it'll take up to 60 seconds before the next node is tried.
This is substantially longer than the clients will wait for a response,
causing an error to be returned to the user.

Setting this timeout to 1 second allows for failover to happen within
the client timeout. This also matches the value used in our app proxy
vhosts[2]. A healthy server should have no trouble establishing a TCP
connection in less than a second.

[1]http://nginx.org/en/docs/http/ngx_http_proxy_module.html#proxy_connect_timeout
[2]https://github.gds/gds/puppet/blob/master/modules/nginx/templates/proxy-vhost.conf#L49