Move traffic replay target config to Hiera
...ready to support traffic replay for the draft machines in future. Do
this now as I'm about to refactor this module further to support traffic
replay to the new Carrenza Staging environment.

Specify the target of the traffic replay in Hiera rather in the Puppet
code, which the previous implementation had avoided (see the removed
comment in the `router::gor` class in this commit's diff).

While hardcoding those values made sense at the time, there's no
simple alternative to moving these to Hiera to support traffic replay
for the draft machines, since the `draft_cache` machine definition
includes the `cache` machine definition.