Remove the performanceplatform_collector class
This is providing no value as it hasn't stopped us being able to deploy
the collectors using capastrano and doesn't include any other settings.

We had issues performing the creation of the /data/apps directory as
capastrano had been run before this class was deployed with the correct
app name. It turns out the directory structure of capastrano differs
from that set up by puppet so it will fail when trying to ensure an app
directory created by capastrano.