Fix signon hostnames in Jenkins after data sync
Once the data sync has run from production to staging we need to fix
the hostnames that are stored in the Signon database.

Signon has a rake task to do this that was added in
https://github.com/alphagov/signonotron2/pull/377

This commit adds 2 optional parameters to the Data_Sync_Complete
Jenkins job. If those parameters are specified then the job will
run the rake task against signon.

We can add the hieradata to the `staging` file rather than the
`staging_migration` file because Jenkins jobs are only managed
on machines which have been migrated.