Add 'ssh-rsa' prefix to Jenkins SSH public keys
Before this change, it was expected that the 'ssh-rsa ' prefix be
included in Hiera data.

Presumably because we specify SSH public keys elsewhere in Hiera data
without the 'ssh-rsa ' prefix [1], this was omitted from the Hiera
data provided for the new Carrenza Staging and Production environments.

Since we're already hard-coding the SSH cipher used by specifying the
SSH public key filename as `id_rsa.pub`, it makes sense just to prepend
'ssh-rsa ' to the public key pulled from Hiera data.

[1]: https://github.gds/gds/puppet/blob/40d99e78c9e0f754ba53159ac2ebbccd4e71b43e/modules/govuk/manifests/deploy/setup.pp#L82-L88