Allow overriding wildcard_alphagov_mgmt_crt
We're going to get rid of the wildcard_alphagov_mgmt certificate
and just have a wildcard_alphagov certificate.

In order to do this, we need to allow the wildcard_alphagov_mgmt_crt
key to be overriden by setting it to UNSET in hieradata.

This code will be revisited soon after the migration is complete.