Rename fqdn_underscore to fqdn_metrics
Rename the `fqdn_underscore` function and change its behaviour.

Previously, it replaced all periods found in a fully-qualified domain
name with underscores.

Additionally, it now also strips the `_publishing_service_gov_uk` suffix
from the fully-qualified domain name.

Do this because our machine hostnames were exceeding the 63-byte limit
that collectd enforces on its host field. By stripping this suffix, we
can ensure that the hostnames that appear in Graphite do not exceed the
63-byte limit, and it's easy for a human to predict what the hostname
will look like.

Rename the function to better reflect the new behaviour.

I wanted to avoid hard-coding the suffix `_publishing_service_gov_uk` in
the function, but that's fine for now. We can fix that later when we're
under less time pressure to migrate to Carrenza.