Allow apps to use an external Licensify service
Most of the GOV.UK stack is migrating from Skyscape to Carrenza.
Licensify and EFG are going to be left behind in Skyscape.

Some apps, like `contentapi`, connect to Licensify using
`Plek.current.find('licensify')`. In order for these apps to
keep working on Carrenza we need to:

(1) Set up a host entry which matches the SSL certificate that's
    available in Skyscape production
(2) Tell Plek that the location of Licensify has changed