Send machine email to a Carrenza-specific list
Send machine email (usually from cron jobs) to a new mailing list I've
just created for environments in Carrenza.

Note that all environments will send to this list, but we don't check
this mailing list enough that that's likely to be an issue.

We should also migrate the remaining things that send machine email to
use passive Icinga checks or log to Kibana.