Configure EFG domains from hieradata
This commit removes the global hiera lookup for the domain name
that EFG is served on and allows it to be set using class parameters.

As well as removing the hiera lookup, this means that we can remove the
host entry on the staging and production Carrenza machines which
will allow our smoke tests to work.