Stop traffic replay on Skyscape Staging
Stop replaying traffic against Skyscape Staging because we're
decommissioning most that environment.

Traffic replay did not work correctly against both of these targets
(Skyscape Staging and Carrenza Staging); it appears the vShield Edge
Gateway was a bottleneck; possibly the number of ephemeral sockets
available on the vSE meant that gor ran out of file descriptors when
targeting both environments, ostensibly because an unusually large
number of requests were in flight.