Allow process type to be set in procfile::worker
If an application has more than one kind of worker then we will want to
define which process type will be pulled out of the Procfile and
executed. You can find a little more information about process types in:

https://devcenter.heroku.com/articles/procfile#declaring-process-types

This extends the current govuk::procfile::worker so that you can pass
through a process_type and it will execute that rather than the default
'worker'.

Signed-off-by: Tom Booth <tombooth@gmail.com>