Change email to notify for complaint report Jenkins job
I have been getting emails when this job fails (which happens frequently)
but I'm no longer on the team which owns the complaint rate report. This
replaces my email address with Jake's; he's happy with this for now. Notifying
a single person isn't a good long-term solution, but at the moment no action
needs to be taken when the job fails, as long as it's succeeding occasionally.
Core Formats are planning to do some work on this at some point to make it
more reliable, and will revisit the notification then.